MARSdata S4 object

S4 class that organizes the various data inputs for the MARS model. MARSdata simply inherits the slots from 6 component classes: Dmodel, Dstock, Dfishery, Dsurvey DCKMR, and Dtag, where the D- prefix denotes an object for model data.

Details

For convenience, most arrays and matrices have the associated dimensions in the variable name. For example, Cobs_ymfr represents observed catch with the dimension following the underscore, following this template:

y	Year

m	Season

a	Age

r	Region

f	Fishery

i	Index

s	Stock

Slots

Dmodel

Class Dmodel containing parameters for model structure (number of years, ages, etc.)

Dstock

Class Dstock containing stock parameters (growth, natural mortality, etc.)

Dfishery

Class Dfishery containing fishery data (catch, size and stock composition, etc.)

Dsurvey

Class Dsurvey containing survey data (indices of abundance)

DCKMR

Class DCKMR containing genetic close-kin data

Dtag

Class Dtag containing tagging data

Misc

List for miscellaneous inputs as needed

Slots inherited from Dmodel

ny

Integer, number of years

nm

Integer, number of seasons

na

Integer, number of ages

nl

Integer, number of length bins. Set to zero if lengths are not modeled.

nr

Integer, number of spatial regions

ns

Integer, number of stocks

lbin

Vector of lower boundary of length bins. Length nl + 1

lmid

Vector of midpoint of length bins. Length nl

Fmax

Numeric, maximum allowable instantaneous fishing mortality rate (units of per season). Defaults to 3.

nitF

Integer, number of iterations to solve Baranov catch equation from observed catch. Defaults to 5.

dist_type

Character. Whether to estimate seasonal stock distribution with a movement matrix "mov" or re-distributing total abundance by a distribution vector "dist"

y_phi

Integer, the year from which to obtain values of natural mortality and fecundity for the unfished stock-recruit replacement line (phi). Relevant if natural mortality or fecundity are time-varying. Defaults to 1.

Slots inherited from Dstock

len_ymas

Length-at-age. Only needed if Dmodel@nl > 0.

sdlen_ymas

Standard deviation in length-at-age

LAK_ymals

Length-at-age probability array. If empty, values will be calculated in check_data (see calc_LAK).

mat_yas

Proportion mature by age class

fec_yas

Fecundity, i.e., spawning output, of mature animals

swt_ymas

Stock weight-at-age

Md_yas

Natural morality. Ignored if M is estimated.

m_spawn

Integer, season of spawning

m_rec

Integer, season of recruitment

SRR_s

Character vector of stock-recruit relationship by stock. See calc_recruitment for options.

delta_s

Fraction of season that elapses when spawning occurs

natal_rs

Boolean matrix that indicates whether stock s spawns when in region r at time of spawning.

Slots inherited from Dfishery

nf

Integer, number of fleets

Cobs_ymfr

Total fishery catch

fwt_yamfs

Fishery weight at age. Set to 1 when fleet catch is in units of abundance. Set to stock weight at age by default.

CAAobs_ymafr

Fishery catch at age composition

CALobs_ymlfr

Fishery catch at length composition

fcomp_like

Character, likelihood for the fishery composition data. See type argument in like_comp for options

CAAN_ymfr

Sample size of the catch at age vector by season if using the multinomial or Dirichlet-multinomial likelihoods

CALN_ymfr

Sample size of the catch at length vector by season if using the multinomial or Dirichlet-multinomial likelihoods

CAAtheta_f

Catch at age dispersion parameter if using the Dirichlet-multinomial likelihood. Default set to 1.

CALtheta_f

Catch at length dispersion parameter if using the Dirichlet-multinomial likelihood. Default set to 1.

sel_block_yf

Index of dummy fleets to model time blocks of selectivity

sel_f

Character matrix of functional form for selectivity

SC_ymafrs

Stock composition data.

SC_aa

Boolean matrix that groups together age classes for the stock composition data

SC_ff

Boolean matrix that groups together fleets for the stock composition data

SC_like

Character, likelihood for the stock composition data. See type argument of like_comp for options

SCN_ymafr

Sample size of the stock composition vector if using the multinomial or Dirichlet-multinomial likelihoods

SCtheta_f

Stock composition dispersion parameter if using the Dirichlet-multinomial likelihood. Default set to 1.

SCstdev_f

Stock composition standard deviation if using the lognormal likelihood. Default set to 0.1.

Slots inherited from Dsurvey

ni

Integer, number of indices of abundance. Zero is possible.

Iobs_ymi

Observed indices

unit_i

Character vector, units of the index. Set to "B" to use stock weight at age (default) or "N" for abundance (numbers).

IAAobs_ymai

Survey age composition

IALobs_ymli

Survey length composition

icomp_like

Character, likelihood for the composition data. See like_comp for options

IAAN_ymi

Sample size of the index age composition by season if using the multinomial or Dirichlet-multinomial likelihoods

IALN_ymi

Sample size of the index length composition by season if using the multinomial or Dirichlet-multinomial likelihoods

IAAtheta_i

Index age composition dispersion parameter if using the Dirichlet-multinomial likelihood

IALtheta_i

Index length composition dispersion parameter if using the Dirichlet-multinomial likelihood

samp_irs

Boolean array that specifies the regions and stocks sampled by the index. samp[i, r, s] indicates whether index i operates in region r and catches stock s.

sel_i

Character matrix of functional forms for selectivity

delta_i

The elapsed fraction of time in the seasonal time step when the index samples the population.

Slots inherited from DCKMR

POP_s

A list by stock of data frames for parent-offspring pairs. Each row in the data frame corresponds to a "sampling unit" defined by the columns:

a	Capture year of parent

t	Age at capture of parent

y	Birth year of offspring

n	Number of pairwise comparisons

m	Number of POPs

HSP_s

A list by stock of data frames for half-sibling pairs. Each row in the data frame corresponds to a "sampling unit" defined by the columns:

yi	Birth year of older sibling

yj	Birth year of younger sibling

n	Number of pairwise comparisons

m	Number of HSPs

CKMR_like

Character, likelihood for the POP and HSP sampling units. See like_CKMR for options.

See also