Plotting functions for data in MARS model

A set of functions to plot data variables and predicted values (catch, age composition, etc.)

Usage

plot_catch(fit, f = 1, by = c("region", "stock"), prop = FALSE, annual = FALSE)

plot_index(fit, i = 1, zoom = FALSE)

plot_CAA(fit, f = 1, r = 1)

plot_CAL(fit, f = 1, r = 1)

plot_IAA(fit, i = 1)

plot_IAL(fit, i = 1)

plot_SC(fit, ff = 1, aa = 1, r = 1, prop = FALSE)

plot_tagmov(fit, s = 1, yy = 1, aa = 1)

Arguments

fit

MARSassess object returned by fit_MARS()

f

Integer, indexes the fleet

by

Character to indicate dimension for multivariate plots

prop

Logical, whether to plot proportions (TRUE) or absolute numbers

annual

Logical, whether to plot annual values (summed over seasons)

i

Integer, indexes the survey

zoom

Logical, for plot_index(). If TRUE, plots a subset of years with observed data points. Otherwise, plots predicted values over all model years.

r

Integer, indexes the region

ff

Integer, indexes the aggregate fleet (for stock composition data)

aa

Integer, indexes the aggregate age class (for stock composition and tag data)

s

Integer, indexes the stock

yy

Integer, indexes the aggregate years (for the tag data)

Value

Various base graphics plots

Details

plot_catch plots the fishery catch by stock or region (either whole numbers or proportions)

plot_index plots indices of abundance

plot_CAA plots the fishery catch at age

plot_CAL plots the catch at length

plot_IAA plots the index age composition

plot_IAL plots the index length composition

plot_SC plots the stock composition

plot_tagmov plots the tag movements