Plotting functions for fitted MARS model

A set of functions to plot state variables (biomass, recruitment time series, etc.)

Usage

plot_S(fit, by = c("total", "stock", "region"), r, s, prop = FALSE)

plot_B(fit, by = c("total", "stock", "region"), r, s, prop = FALSE)

plot_R(fit, s)

plot_SRR(fit, s = 1, phi = TRUE)

plot_Rdev(fit, s = 1, log = TRUE)

plot_Fstock(fit, s, by = c("annual", "season"))

plot_self(fit, f = 1, type = c("length", "age"))

plot_seli(fit, i = 1)

plot_selstock(
  fit,
  s = 1,
  by = c("annual", "season"),
  plot2d = c("contour", "filled.contour"),
  ...
)

plot_N(fit, m = 1, r, s = 1, plot2d = c("contour", "filled.contour"), ...)

plot_V(fit, f = 1, by = c("stock", "region"), prop = FALSE)

plot_Ffleet(fit, f = 1)

plot_mov(fit, s = 1, y, a)

plot_recdist(fit)

Arguments

fit

MARSassess object returned by fit_MARS()

by

Character to indicate whether to calculate selectivity from F per year or per season

r

Integer for the corresponding region

s

Integer for the corresponding stock

prop

Logical, whether to plot proportions (TRUE) or absolute numbers

phi

Logical, whether to plot unfished replacement line

log

Logical, whether to plot the natural logarithm of the response variable

f

Integer for the corresponding fleet

type

For plot_self, indicates whether to plot the selectivity by age or length.

i

Integer for the corresponding survey

plot2d

Character, plotting function for either a contour() or filled.contour() plot

...

Other argument to the base graphics function

m

Integer for the corresponding season

y

Integer, year for plotting the movement matrix

a

Integer, corresponding age for plotting the movement matrix

Value

Various base graphics plots

Details

plot_S plots spawning output by stock or region (either whole numbers or proportions for the latter)

plot_B plots total biomass by stock or region (either whole numbers or proportions for the latter)

plot_R plots recruitment by stock

plot_SRR plots the stock-recruitment relationship and history (realized recruitment) by stock

plot_Rdev plots recruitment deviations by stock

plot_Fstock plots apical instantaneous fishing mortality (per year or per season) by stock

plot_self plots fishery selectivity

plot_seli plots index selectivity

plot_selstock plots the realized selectivity from total catch and total abundance at age

plot_N reports total abundance at age

plot_V plots vulnerable biomass, availability to the fishery

plot_Ffleet plots apical instantaneous fishing mortality (per season) by fleet

plot_mov plots movement matrices and the corresponding equilibrium distribution in multi-area models

plot_recdist plots the distribution of recruitment for each stock