Dmodel S4 object

S4 class that organizes the various data inputs for the MARS model. MARSdata simply inherits the slots from 6 component classes: Dmodel, Dstock, Dfishery, Dsurvey DCKMR, and Dtag, where the D- prefix denotes an object for model data.

Details

For convenience, most arrays and matrices have the associated dimensions in the variable name. For example, Cobs_ymfr represents observed catch with the dimension following the underscore, following this template:

y	Year

m	Season

a	Age

r	Region

f	Fishery

i	Index

s	Stock

Slots inherited from Dmodel

ny

Integer, number of years

nm

Integer, number of seasons

na

Integer, number of ages

nl

Integer, number of length bins. Set to zero if lengths are not modeled.

nr

Integer, number of spatial regions

ns

Integer, number of stocks

lbin

Vector of lower boundary of length bins. Length nl + 1

lmid

Vector of midpoint of length bins. Length nl

Fmax

Numeric, maximum allowable instantaneous fishing mortality rate (units of per season). Defaults to 3.

nitF

Integer, number of iterations to solve Baranov catch equation from observed catch. Defaults to 5.

dist_type

Character. Whether to estimate seasonal stock distribution with a movement matrix "mov" or re-distributing total abundance by a distribution vector "dist"

y_phi

Integer, the year from which to obtain values of natural mortality and fecundity for the unfished stock-recruit replacement line (phi). Relevant if natural mortality or fecundity are time-varying. Defaults to 1.

See also