Dsurvey S4 object

S4 class that organizes the various data inputs for the MARS model. MARSdata simply inherits the slots from 6 component classes: Dmodel, Dstock, Dfishery, Dsurvey DCKMR, and Dtag, where the D- prefix denotes an object for model data.

Details

For convenience, most arrays and matrices have the associated dimensions in the variable name. For example, Cobs_ymfr represents observed catch with the dimension following the underscore, following this template:

y	Year

m	Season

a	Age

r	Region

f	Fishery

i	Index

s	Stock

Slots inherited from Dsurvey

ni

Integer, number of indices of abundance. Zero is possible.

Iobs_ymi

Observed indices

Isd_ymi

Lognormal standard deviation of the observed indices

unit_i

Character vector, units of the index. Set to "B" to use stock weight at age (default) or "N" for abundance (numbers).

IAAobs_ymai

Survey age composition

IALobs_ymli

Survey length composition

icomp_like

Character, likelihood for the composition data. See like_comp() for options

IAAN_ymi

Sample size of the index age composition by season if using the multinomial or Dirichlet-multinomial likelihoods

IALN_ymi

Sample size of the index length composition by season if using the multinomial or Dirichlet-multinomial likelihoods

IAAtheta_i

Index age composition dispersion parameter if using the Dirichlet-multinomial likelihood

IALtheta_i

Index length composition dispersion parameter if using the Dirichlet-multinomial likelihood

samp_irs

Boolean array that specifies the regions and stocks sampled by the index. samp[i, r, s] indicates whether index i operates in region r and catches stock s.

sel_i

Character matrix of functional forms for selectivity. See "type" argument in conv_selpar() for options.

delta_i

The elapsed fraction of time in the seasonal time step when the index samples the population.

See also