Calculate index at age

For indices of abundance, the function calculates the numbers vulnerable to the survey.

calc_index(
  N,
  Z,
  sel,
  na = dim(N)[1],
  ni = dim(sel)[2],
  ns = dim(N)[3],
  samp = array(1, c(ni, na, ns)),
  delta = rep(0, ni)
)

Arguments

N

Stock abundance at the beginning of the time step. Array [a, r, s]

Z

Instantaneous total mortality. Array [a, r, s]

sel

Index selectivity. Array [a, i, s]

na

Integer, number of age classes

ni

Integer, number of indices

ns

Integer, number of stocks

samp

Boolean indicates which regions and stocks are sampled by the index. Array [i, r, s]

delta

Fraction of time step when the index samples the population. Vector by i

Value

Index at age. Array [a, i, s]

Details

The index is calculated as $$ I_{a,i,s} = v_{a,i,s} \sum_r N_{a,r,s} \exp(-\delta_i Z_{a,r,s}) \times \mathbb{1}(r \in R_i) \mathbb{1}(s \in S_i) $$

where \(R_i\) and \(S_i\) denote the regions and stocks, respectively, sampled by index \(i\). For example, \(R_2 = 1\) denotes that the second index of abundance only samples region 1. These are informed by array samp where samp[i, r, s] = 1 indicates that stock s in region r is sampled by index i.